backend/requirements.txt
asgiref==3.7.2
astroid==3.0.1
colorama==0.4.6
dill==0.3.7
Django==4.2.7
django-cors-headers==4.3.1
django-filter==23.4
djangorestframework==3.14.0
environ==1.0
isort==5.12.0
Markdown==3.5.1
mccabe==0.7.0
platformdirs==4.0.0
pylint==3.0.2
pylint-django==2.5.5
pylint-plugin-utils==0.8.2
python-version==0.0.2
pytz==2023.3.post1
sqlparse==0.4.4
tomlkit==0.12.3
tzdata==2023.3
